assignment-3: rearranged delete and update handling
