+++ tv-shows-app/src/components/feature/review/ReviewItem/ReviewItem.tsx
@@ -7,7 +7,8 @@ import { swrKeys } from "@/fetchers/swrKeys";
 import { deleteReview } from "@/fetchers/mutators";
 import { authFetcher } from "@/fetchers/fetcher";
 import { IUser } from "@/typings/user";
-import ReviewUpdate from "../ReviewUpdate/ReviewUpdate";
+import ReviewUpdate from "./components/ReviewUpdate/ReviewUpdate";
+import ReviewDeleteButton from "./components/ReviewDeleteButton/ReviewDeleteButton";
 
 export interface IReviewItemProps {
    review: IReview
@@ -15,19 +16,6 @@ export interface IReviewItemProps {
 
 export default function ReviewItem({review} : IReviewItemProps) {
    const {data} = useSWR(swrKeys.me, authFetcher<{user: IUser}>);
-
-   const { trigger } = useSWRMutation(swrKeys.reviews(`/${review.id}`), deleteReview, {
-      onSuccess: () => {
-         mutate(swrKeys.getReviews(review.show_id))
-      }
-   });
-
-   const removeReview = async () => {
-      try {
-         await trigger();
-      }
-      catch(error) {}
-   }
 
    return <Card padding={1} backgroundColor={'lightblue'} color={'white'}>
       <Flex direction={'column'} gap = {1}>
@@ -40,7 +28,7 @@ export default function ReviewItem({review} : IReviewItemProps) {
          </Flex>
          <Text data-testid="text">{review.comment}</Text>
          <ReviewStarsInput label = {`${review.rating} / 5`} value={review.rating} onChange={() => {return;}}></ReviewStarsInput>
-         {data?.user.email === review.user?.email && <Button width={'30%'} onClick={() => {removeReview()}}>Remove</Button>}
+         {data?.user.email === review.user?.email && <ReviewDeleteButton review={review}/>}
       </Flex>
    </Card>
 }
